+++ electron/preload.js
@@ -3,7 +3,7 @@ const { contextBridge, ipcRenderer, webFrame } = require('electron');
 contextBridge.exposeInMainWorld('moor', {
   zoomBy: (delta) => webFrame.setZoomLevel(webFrame.getZoomLevel() + delta),
   zoomReset: () => webFrame.setZoomLevel(0),
-  readFile: (filePath) => ipcRenderer.invoke('read-file', { filePath }),
+  readFile: (filePath, asText = false) => ipcRenderer.invoke('read-file', { filePath, asText }),
   readFileAsDataUrl: (filePath) => ipcRenderer.invoke('read-file-as-data-url', { filePath }),
   compareDirectories: (leftPath, rightPath) => ipcRenderer.invoke('compare-directories', { leftPath, rightPath }),
   onInitialPaths: (callback) => ipcRenderer.once('initial-paths', (_event, data) => callback(data)),